Adding a pre-end of execution event, for flushing before MPI_Finalize.
Former-commit-id: d7ae72ecf4889e44e8576cb25ef62b708c5c5bd5

+++ include/Profile/TauUtil.h
@@ -53,16 +53,16 @@ void *Tau_util_calloc(size_t size, const char *file, int line);
 #ifdef __cplusplus
 extern "C" {
 #endif /* __cplusplus */
-void Tau_plugin_sendmsg(int type, int destination, int length, int remoteid);
-void Tau_plugin_recvmsg(int type, int source, int length, int remoteid);
+void Tau_plugin_sendmsg(long unsigned int type, long unsigned int destination, long unsigned int length, long unsigned int remoteid);
+void Tau_plugin_recvmsg(long unsigned int type, long unsigned int source, long unsigned int length, long unsigned int remoteid);
 #ifdef __cplusplus
 }
 #endif /* __cplusplus */
 
 // Use a macro so we can compile it out if too much overhead.
 
-#define TAU_PLUGIN_SENDMSG(t, d, l, r) Tau_plugin_sendmsg(t, d, l, r)
-#define TAU_PLUGIN_RECVMSG(t, s, l, r) Tau_plugin_recvmsg(t, s, l, r)
+#define TAU_PLUGIN_SENDMSG(t, d, l, r) Tau_plugin_sendmsg((long unsigned int)t, (long unsigned int)d, (long unsigned int)l, (long unsigned int)r)
+#define TAU_PLUGIN_RECVMSG(t, s, l, r) Tau_plugin_recvmsg((long unsigned int)t, (long unsigned int)s, (long unsigned int)l, (long unsigned int)r)
 
 /* The following macros help create a local array and assign to elements of 
    the local C array, values from Fortran array after conversion using f2c
